requirements.txt
altgraph==0.14
certifi==2017.11.5
chardet==3.0.4
click==6.7
EasyProcess==0.2.3
Flask==0.12.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
future==0.16.0
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
macholib==1.8
MarkupSafe==1.0
olefile==0.44
pefile==2017.11.5
Pillow==4.3.0
pygeoip==0.3.2
PyInstaller==3.3
pyscreenshot==0.4.2
requests==2.18.4
SQLAlchemy==1.1.15
urllib3==1.22
Werkzeug==0.12.2

